GH-822 using review controller
